requirements.txt
anthropic>=0.40.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
requests>=2.32.0
beautifulsoup4>=4.12.0
notion-client>=2.2.1
click>=8.1.0
python-dotenv>=1.0.0
